# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Telemetria GPS",
        "description": "API dedicada à extração e consumo de dados de rastreamento de frotas.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/esp32/telemetry": {
            "post": {
                "tags": [
                    "ESP32"
                ],
                "summary": "Recebe e processa dados de telemetria de uma placa ESP32",
                "description": "Endpoint público para ingestão de dados. Cria o dispositivo automaticamente se não existir (firstOrCreate por identificador/MAC).",
                "operationId": "4e11f793b4dc03072b6fca9b20f7c082",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "identificador"
                                ],
                                "properties": {
                                    "identificador": {
                                        "description": "MAC Address ou ID único do chip",
                                        "type": "string",
                                        "example": "AA:BB:CC:DD:EE:FF"
                                    },
                                    "nome": {
                                        "description": "Nome amigável do dispositivo",
                                        "type": "string",
                                        "example": "ESP32-Caminhão-01"
                                    },
                                    "lat": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -23.55052
                                    },
                                    "lon": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -46.6333
                                    },
                                    "bateria": {
                                        "description": "Voltagem da bateria (V)",
                                        "type": "number",
                                        "format": "float",
                                        "example": 3.85
                                    },
                                    "temp": {
                                        "description": "Temperatura em °C",
                                        "type": "number",
                                        "format": "float",
                                        "example": 34.5
                                    },
                                    "vel": {
                                        "description": "Velocidade em km/h",
                                        "type": "integer",
                                        "example": 60
                                    },
                                    "timestamp": {
                                        "description": "Data/hora do dispositivo (opcional, usa now() se omitido)",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-05-03T20:00:00Z"
                                    },
                                    "extra": {
                                        "description": "Payload extra em JSON livre",
                                        "type": "object",
                                        "example": {
                                            "sinal_gsm": -75
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Telemetria processada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Dados inválidos",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/esp32/{identificador}/historico": {
            "get": {
                "tags": [
                    "ESP32"
                ],
                "summary": "Retorna o histórico de telemetria de um dispositivo ESP32 por período",
                "operationId": "53e2ebfd3a88b7dfb2d764be24071e78",
                "parameters": [
                    {
                        "name": "identificador",
                        "in": "path",
                        "description": "MAC Address ou ID do dispositivo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "data_inicio",
                        "in": "query",
                        "description": "Data de início (Y-m-d H:i:s)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "data_fim",
                        "in": "query",
                        "description": "Data final (Y-m-d H:i:s)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "por_pagina",
                        "in": "query",
                        "description": "Registros por página (padrão 100)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Histórico paginado"
                    },
                    "404": {
                        "description": "Dispositivo não encontrado"
                    },
                    "422": {
                        "description": "Parâmetros inválidos"
                    }
                }
            }
        },
        "/api/v1/esp32/{identificador}/ultima": {
            "get": {
                "tags": [
                    "ESP32"
                ],
                "summary": "Retorna a última telemetria recebida de um dispositivo ESP32",
                "operationId": "bf971a537e938c54b9614299764a3465",
                "parameters": [
                    {
                        "name": "identificador",
                        "in": "path",
                        "description": "MAC Address ou ID do dispositivo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Última telemetria do dispositivo"
                    },
                    "404": {
                        "description": "Dispositivo ou telemetria não encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}